Issue-#10: WIP
